+++ lib/src/models/documents/delta_x.dart
@@ -1,9 +1,8 @@
-import 'package:html2md/html2md.dart' as html2md;
+import 'package:flutter_quill_delta_from_html/flutter_quill_delta_from_html.dart';
 import 'package:markdown/markdown.dart' as md;
 import 'package:meta/meta.dart';
 import '../../../markdown_quill.dart';
 import '../../../quill_delta.dart';
-import '../../utils/delta_x_utils.dart';
 
 @immutable
 @experimental
@@ -15,35 +14,37 @@ class DeltaX {
   /// This api is **experimental** and designed to be used **internally** and shouldn't
   /// used for **production applications**.
   @experimental
+  @Deprecated(
+    '''
+    The experimental support for Markdown conversion has been dropped and will be removed in future releases, 
+    consider using alternatives such as https://pub.dev/packages/markdown_quill
+    ''',
+  )
   static Delta fromMarkdown(String markdownText) {
-    final mdDocument = md.Document(
-      encodeHtml: false,
-      inlineSyntaxes: [UnderlineSyntax(), VideoSyntax()],
-    );
+    final mdDocument = md.Document(encodeHtml: false);
     final mdToDelta = MarkdownToDelta(markdownDocument: mdDocument);
     return mdToDelta.convert(markdownText);
   }
 
   /// Convert the HTML Raw string to [Delta]
-  ///
-  /// It will run using the following steps:
-  ///
-  /// 1. Convert the html to markdown string using `html2md` package
-  /// 2. Convert the markdown string to [Delta] using [fromMarkdown]
-  ///
-  /// This api is **experimental** and designed to be used **internally** and shouldn't
-  /// used for **production applications**.
-  ///
   @experimental
-  static Delta fromHtml(String htmlText) {
-    final markdownText = html2md.convert(
-      htmlText,
-      rules: [underlineRule, videoRule],
-      styleOptions: {'emDelimiter': '*'},
-    ).replaceAll(
-      'unsafe:',
-      '',
-    );
-    return fromMarkdown(markdownText);
+  @Deprecated(
+    '''
+    The experimental support for HTML conversion has been dropped and will be removed in future releases, 
+    consider using alternatives such as https://pub.dev/packages/flutter_quill_delta_from_html
+    ''',
+  )
+  static Delta fromHtml(
+    String htmlText, {
+    @Deprecated(
+      '''
+    The experimental support for HTML conversion has been dropped and will be removed in future releases, 
+    consider using alternatives such as https://pub.dev/packages/flutter_quill_delta_from_html
+    ''',
+    )
+    List<CustomHtmlPart>? customBlocks,
+  }) {
+    final htmlToDelta = HtmlToDelta(customBlocks: customBlocks);
+    return htmlToDelta.convert(htmlText);
   }
 }