+++ lib/src/widgets/quill/quill_controller.dart
@@ -568,7 +568,7 @@ class QuillController extends ChangeNotifier {
 
   /// Return true if can paste using HTML
   Future<bool> _pasteHTML() async {
-    final clipboardService = ClipboardServiceProvider.instacne;
+    final clipboardService = ClipboardServiceProvider.instance;
 
     Future<String?> getHTML() async {
       if (await clipboardService.canProvideHtmlTextFromFile()) {
@@ -583,6 +583,7 @@ class QuillController extends ChangeNotifier {
     final htmlText = await getHTML();
     if (htmlText != null) {
       final htmlBody = html_parser.parse(htmlText).body?.outerHtml;
+      // ignore: deprecated_member_use_from_same_package
       final deltaFromClipboard = DeltaX.fromHtml(htmlBody ?? htmlText);
 
       _pasteUsingDelta(deltaFromClipboard);
@@ -594,7 +595,7 @@ class QuillController extends ChangeNotifier {
 
   /// Return true if can paste using Markdown
   Future<bool> _pasteMarkdown() async {
-    final clipboardService = ClipboardServiceProvider.instacne;
+    final clipboardService = ClipboardServiceProvider.instance;
 
     Future<String?> getMarkdown() async {
       if (await clipboardService.canProvideMarkdownTextFromFile()) {
@@ -608,6 +609,7 @@ class QuillController extends ChangeNotifier {
 
     final markdownText = await getMarkdown();
     if (markdownText != null) {
+      // ignore: deprecated_member_use_from_same_package
       final deltaFromClipboard = DeltaX.fromMarkdown(markdownText);
 
       _pasteUsingDelta(deltaFromClipboard);