+++ lib/src/widgets/toolbar/simple_toolbar.dart
@@ -31,6 +31,7 @@ class QuillSimpleToolbar extends StatelessWidget
           configurations.showItalicButton ||
           configurations.showSmallButton ||
           configurations.showUnderLineButton ||
+          configurations.showLineHeightButton ||
           configurations.showStrikeThrough ||
           configurations.showInlineCode ||
           configurations.showColorButton ||
@@ -195,6 +196,12 @@ class QuillSimpleToolbar extends StatelessWidget
                 isButtonGroupShown[4] ||
                 isButtonGroupShown[5]))
           divider,
+        if (configurations.showLineHeightButton)
+          QuillToolbarSelectLineHeightStyleDropdownButton(
+            controller: globalController,
+            options: toolbarConfigurations
+                .buttonOptions.selectLineHeightStyleDropdownButton,
+          ),
         if (configurations.showHeaderStyle) ...[
           if (configurations.headerStyleType.isOriginal)
             QuillToolbarSelectHeaderStyleDropdownButton(